+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @apps/demo
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [3f2798c]
+  - react-router-xray-react@2.0.9
+
 ## 0.1.9
 
 ### Patch Changes